+++ src/main/java/com/RecipeCode/teamproject/reci/feed/recipeslikes/controller/RecipesLikesController.java
@@ -5,8 +5,11 @@ import com.RecipeCode.teamproject.reci.feed.recipeslikes.dto.RecipesLikesDto;
 import com.RecipeCode.teamproject.reci.feed.recipeslikes.service.RecipesLikesService;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.log4j.Log4j2;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -16,34 +19,24 @@ import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequiredArgsConstructor
-@RequestMapping("/recipes")
+@RequestMapping("/api/recipes")
+@Log4j2
 public class RecipesLikesController {
 
     private final RecipesLikesService recipesLikesService;
 
     // 좋아요 토글
 
-//    @PostMapping("/{uuid}/like")
-//    public ResponseEntity<RecipesLikesDto> toggleLike(
-//            @PathVariable String uuid,
-//            @AuthenticationPrincipal SecurityUserDto user){
-//        if (user == null) {
-//            throw new RuntimeException("No user logged in");
-//        }
-//        RecipesLikesDto dto = recipesLikesService.toggleLike(uuid, user.getUsername());
-//        return ResponseEntity.ok(dto);
-//    }
-
-    // 세션 인증 방어 테스트 컨트롤러
     @PostMapping("/{uuid}/like")
     public ResponseEntity<RecipesLikesDto> toggleLike(
             @PathVariable String uuid,
-            HttpSession session){
-        String userEmail = (String) session.getAttribute("userEmail");
-        if (userEmail == null) throw new RuntimeException("No user logged in");
-
-        RecipesLikesDto dto = recipesLikesService.toggleLike(uuid, userEmail);
+            @AuthenticationPrincipal SecurityUserDto user){
+        if (user == null) {
+            throw new RuntimeException("No user logged in");
+        }
+        RecipesLikesDto dto = recipesLikesService.toggleLike(uuid, user.getUsername());
         return ResponseEntity.ok(dto);
     }
+
 
 }
