+++ src/main/java/com/RecipeCode/teamproject/reci/mypage/controller/MyPageRestController.java
@@ -1,7 +1,6 @@
 package com.RecipeCode.teamproject.reci.mypage.controller;
 
 import com.RecipeCode.teamproject.reci.auth.dto.SecurityUserDto;
-import com.RecipeCode.teamproject.reci.auth.entity.Member;
 import com.RecipeCode.teamproject.reci.feed.recipes.dto.RecipesDto;
 import com.RecipeCode.teamproject.reci.mypage.service.MyPageService;
 import lombok.RequiredArgsConstructor;
@@ -11,7 +10,6 @@ import org.springframework.data.domain.Slice;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.web.PageableDefault;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
-import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -20,7 +18,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/mypage")
 @RequiredArgsConstructor
 @Log4j2
-public class MyPageController {
+public class MyPageRestController {
     private final MyPageService myPageService;
 
     @GetMapping("/my-recipes")
