+++ src/main/java/com/RecipeCode/teamproject/reci/mypage/controller/MyPageViewController.java
@@ -1,19 +1,43 @@
 package com.RecipeCode.teamproject.reci.mypage.controller;
 
+import com.RecipeCode.teamproject.common.MapStruct;
+import com.RecipeCode.teamproject.reci.auth.dto.MemberDto;
+import com.RecipeCode.teamproject.reci.auth.dto.SecurityUserDto;
+import com.RecipeCode.teamproject.reci.auth.entity.Member;
+import com.RecipeCode.teamproject.reci.auth.repository.MemberRepository;
+import com.RecipeCode.teamproject.reci.auth.service.MemberService;
 import com.RecipeCode.teamproject.reci.mypage.service.MyPageService;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 
 @Controller
 @RequiredArgsConstructor
-@RequestMapping("/mypage")
 public class MyPageViewController {
-    private final MyPageService myPageService;
 
-    @GetMapping("/my-recipes")
-    public String myPageView(){
+    private final MemberService memberService;
+    private final MapStruct mapStruct;
+
+    @GetMapping("/mypage")
+    public String myPageView(@AuthenticationPrincipal SecurityUserDto principal,
+                             Model model) {
+        // 1) 비로그인일 때 로그인 페이지로
+        if (principal == null) {
+            return "redirect:/auth/login";
+        }
+
+        // 2) 로그인 이메일로 Member 조회
+        String email = principal.getUsername(); // SecurityUserDto의 username = userEmail
+        Member user = memberService.getByUserEmail(email);
+        MemberDto memberDto = mapStruct.toDto(user);
+        // 3) 화면에 뿌리기
+        model.addAttribute("user", memberDto);
+
+        // 4) JS에서 필요하면 바로 씀 (예: mypage-feed.js)
+         model.addAttribute("currentUserEmail", memberDto.getUserEmail());
         return "profile/mypage_all";
     }
 }
