+++ src/main/webapp/WEB-INF/views/feed/recipe_details.jsp
@@ -218,8 +218,8 @@
 <script>
     const ctx = "${pageContext.request.contextPath}";
 </script>
-<script src="${ctx}/js/recipe-detail-common.js"></script>
-<script src="${ctx}/js/recipe-details.js"></script>
+<script src="${ctx}/js/recipes/recipe-detail-common.js"></script>
+<script src="${ctx}/js/recipes/recipe-details.js"></script>
 
 
 <%-- TODO: 신고 모달 --%>
@@ -252,5 +252,7 @@
         </form>
     </div>
 </div>
+
+
 </body>
 </html>
