+++ src/main/webapp/js/recipes/recipe-detail-common.js
